drivers/safenet.c: fix ondelay/offdelay handling from int to long (to match strtol() used)

+++ drivers/safenet.c
@@ -60,8 +60,8 @@ static union	{
 	struct safenet		status;
 } ups;
 
-static int ondelay = 1;		/* minutes */
-static int offdelay = 30;	/* seconds */
+static long ondelay = 1;		/* minutes */
+static long offdelay = 30;	/* seconds */
 
 static int safenet_command(const char *command)
 {
@@ -329,8 +329,8 @@ void upsdrv_initinfo(void)
 	dstate_setinfo("ups.model", "%s", ((v = getval("modelname")) != NULL) ? v : "unknown");
 	dstate_setinfo("ups.serial", "%s", ((v = getval("serialnumber")) != NULL) ? v : "unknown");
 
-	dstate_setinfo("ups.delay.start", "%d", 60 * ondelay);
-	dstate_setinfo("ups.delay.shutdown", "%d", offdelay);
+	dstate_setinfo("ups.delay.start", "%ld", 60 * ondelay);
+	dstate_setinfo("ups.delay.shutdown", "%ld", offdelay);
 
 	/*
 	 * These are the instant commands we support.
@@ -509,7 +509,7 @@ void upsdrv_initups(void)
 	}
 
 	if ((ondelay < 0) || (ondelay > 9999)) {
-		fatalx(EXIT_FAILURE, "Start delay '%d' out of range [0..9999]", ondelay);
+		fatalx(EXIT_FAILURE, "Start delay '%ld' out of range [0..9999]", ondelay);
 	}
 
 	val = getval("offdelay");
@@ -518,7 +518,7 @@ void upsdrv_initups(void)
 	}
 
 	if ((offdelay < 0) || (offdelay > 999)) {
-		fatalx(EXIT_FAILURE, "Shutdown delay '%d' out of range [0..999]", offdelay);
+		fatalx(EXIT_FAILURE, "Shutdown delay '%ld' out of range [0..999]", offdelay);
 	}
 }
